splitting off rating key tracking into a separate component
